changed implementation of USE_LOCAL_CC65 to allow building from clean checkout, update README.md
instead of checking if the VAR is EMPTY ie:
 ifeq ($(USE_LOCAL_CC65),"")
(which is also true if the VAR is not set nor passed in),
this patch checks to see if the VAR is set to anything, in which case the LOCAL is used.

Therefore, the default behaviour is to build the submodule.
ie
```
 make                       SUBMODULE
 make USE_LOCAL_CC65=       SUBMODULE
 make USE_LOCAL_CC65=1      LOCAL
```

ref: https://unix.stackexchange.com/question/389717/pass-an-option-to-a-makefile

Note that I tried the:
 ifeq ($(USE_LOCAL_CC65),"1")
but didnt seem to get a match with "make USE_LOCAL_CC65=1"